PR fixes No variable shadowing in WorldModel Function isinstance should be used instead of issubclass wherever possible Item prefixes and suffixes constants should belong to Item Inventory should not be nested in HasInventory Inventory.Command should be Inventory.ItemCommand
